BIG 254: Split lunch/dinner and drinkupload pages into 2
Lunch page: Now cannot change category since it will always be lunch. You can now check a box if an item has multiple prices. If one price, the variable in the formData is price. If two prices are used, the variables in the formData are price_small and price_large, and price is not used. This is done to match the database items structured the same way.

Drink page: You can select boxes based on the number of prices you want. A notice was added to the page about how to correctly change the price of a drink category. Prices are read from the array stored in the database, parsed and split into different variables, and then the final prices are put back into an array to be passed back into the database.